requirements.txt
amqp==5.1.1
asgiref==3.6.0
billiard==3.6.4.0
Brotli==1.0.9
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
crispy-bootstrap5==0.7
cryptography==39.0.0
cssselect2==0.7.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.1.6
#psycopg==3.1.8
psycopg2-binary
django-admin-extra-buttons==1.5.6
django-allauth==0.52.0
django-autoslug==1.9.8
django-countries==7.5.1
django-crispy-forms==2.0
django-import-export==3.0.2
django_debug_toolbar==3.8.1
et-xmlfile==1.1.0
fonttools==4.38.0
html5lib==1.1
idna==3.4
kombu==5.2.4
MarkupPy==1.14
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.0
Pillow==9.4.0
prompt-toolkit==3.0.36
psycopg==3.1.8
psycopg2==2.9.5
pycparser==2.21
pydyf==0.5.0
PyJWT==2.6.0
pyphen==0.13.2
python3-openid==3.2.0
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.3
tablib==3.3.0
tinycss2==1.2.1
typing_extensions==4.4.0
tzdata==2022.7
django-autoslug
Pillow
django-environ
gunicorn
urllib3==1.26.14
vine==5.0.0
wcwidth==0.2.6
weasyprint==58.0
webencodings==0.5.1
xlrd==2.0.1
xlwt==1.3.0
zopfli==0.2.2
